+++ 00-AI/scripts/kb/__init__.py
@@ -7,5 +7,4 @@ from .install import *
 from .intake import *
 from .migrate import *
 from .project import *
-from .release import *
 from .utils import *